nic_router: use increasing src port for new nat
The nat feature of nic_router always re-used src ports that have been
freed recently. From a external server's perspective, if a client dies
and restarts, the new connect would arrive with the src_ip/src_port
combination of the old connection. The server will forcefully reset the
new connection. If that happens a lot, the server may even start to
ignore further connections from this ip/port combination for a while as
a mitigation.

This patch adds a continuous counter feature that makes sure that every
new port allocation will increment and result in an port that hasn't
been used for a long time.

The nat feature of the nic_router is now more in line with RFC 6056
chapter 4.

Issue #4086